LayoutTests/accessibility/aria-listbox-no-selection.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body id="body">
<ul id="select">
    <li>Option 1</li>
    <li>Option 2</li>
    <li>Option 3</li>
</ul>
<p id="description"></p>
<div id="console"></div>
<script>
    description("Tests that native listboxes with no selected children report no selected children.");
    if (window.accessibilityController) {
        var listbox = accessibilityController.accessibleElementById("select");
        shouldBe("listbox.selectedChildrenCount", "0");
        shouldBeTrue("listbox.selectedChildAtIndex(1) == null");
}

</script>

<script src="../resources/js-test-post.js"></script>
</body>
</html>
